refactor(logic): extract world, diplomacy, research, and naval helpers
Made-with: Cursor
