+++ packages/colonizethis_logic/lib/src/economy/worker_economy.dart
@@ -1,14 +1,11 @@
 import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-/// Worker economy primitives (labour and luxuries).
+/// Worker economy primitives (labour, food, and luxuries).
 /// SPEC/program/economy-models.md, SPEC/game/workers-and-population.md.
 ///
 /// Public for use by AI economy planner and production/consumption pipelines.
 
-/// Computes effective available labour for production, capped by luxury
-/// availability per SPEC/program/economy-models.md and
-/// SPEC/game/workers-and-population.md.
 int effectiveLabourForWorkers({
   required WorkerPool workers,
   required Stockpile stockpile,
@@ -21,7 +18,9 @@ int effectiveLabourForWorkers({
 
   final apprenticesWithLuxury = workers.apprentices <= 0
       ? 0
-      : (refinedSugar < workers.apprentices ? refinedSugar : workers.apprentices);
+      : (refinedSugar < workers.apprentices
+          ? refinedSugar
+          : workers.apprentices);
   final journeymenWithLuxury = workers.journeymen <= 0
       ? 0
       : (cigars < workers.journeymen ? cigars : workers.journeymen);
@@ -35,3 +34,76 @@ int effectiveLabourForWorkers({
       mastersWithLuxury * 8;
 }
 
+/// Consumes up to [required] food units (grain/meat) from [stockpile].
+/// Returns a record of (updatedStockpile, unitsConsumed).
+(Stockpile, int) consumeFoodUnits({
+  required Stockpile stockpile,
+  required int required,
+}) {
+  var current = stockpile;
+  var remaining = required;
+  final grainId = CommodityCatalog.grain.id;
+  final meatId = CommodityCatalog.meat.id;
+
+  final grainAvailable = current.quantityOf(grainId);
+  final meatAvailable = current.quantityOf(meatId);
+
+  final grainToUse = remaining <= 0
+      ? 0
+      : remaining <= grainAvailable
+          ? remaining
+          : grainAvailable;
+  if (grainToUse > 0) {
+    current = current.applyDelta(grainId, -grainToUse);
+    remaining -= grainToUse;
+  }
+
+  final meatToUse = remaining <= 0
+      ? 0
+      : remaining <= meatAvailable
+          ? remaining
+          : meatAvailable;
+  if (meatToUse > 0) {
+    current = current.applyDelta(meatId, -meatToUse);
+    remaining -= meatToUse;
+  }
+
+  return (current, required - remaining);
+}
+
+/// Applies tier luxury consumption for trained workers and returns the updated stockpile.
+Stockpile deductLuxuryForWorkers({
+  required Stockpile stockpile,
+  required WorkerPool workers,
+}) {
+  var current = stockpile;
+
+  int _deduct({
+    required int workerCount,
+    required CommodityId luxuryId,
+  }) {
+    if (workerCount <= 0) {
+      return 0;
+    }
+    final available = current.quantityOf(luxuryId);
+    if (available <= 0) {
+      return 0;
+    }
+    final toUse = workerCount < available ? workerCount : available;
+    if (toUse <= 0) {
+      return 0;
+    }
+    current = current.applyDelta(luxuryId, -toUse);
+    return toUse;
+  }
+
+  final refinedSugarId = CommodityCatalog.refinedSugar.id;
+  final cigarsId = CommodityCatalog.cigars.id;
+  final furHatsId = CommodityCatalog.furHats.id;
+
+  _deduct(workerCount: workers.apprentices, luxuryId: refinedSugarId);
+  _deduct(workerCount: workers.journeymen, luxuryId: cigarsId);
+  _deduct(workerCount: workers.masters, luxuryId: furHatsId);
+
+  return current;
+}
